refactor: LottoService를 domain->service 패키지 이동
