+++ src/test/java/domain/LottoServiceTest.java
@@ -8,6 +8,7 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import service.LottoService;
 
 class LottoServiceTest {
